fix: Avoid removing books after backup restore

+++ lib/main.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:path_provider/path_provider.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 import 'core/utils/extensions.dart';
 import 'features/add_new_book/presentation/bloc/book_tags_bloc.dart';
@@ -49,6 +50,10 @@ Future<void> _initDbWithFakes() async {
 }
 
 Future<void> _prepareForBackup() async {
+  final prefs = await SharedPreferences.getInstance();
+  if (prefs.getBool('books-restored') ?? false) {
+    return Future.value();
+  }
   final booksRepo = sl<LocalBooksRepository>();
   await booksRepo.deleteAll();
   var content = await rootBundle.load('assets/backup.json');
